requirements.txt
PyQt6==6.7.1
python-vlc
mutagen==1.47.0
